+++ sklearn/cluster/tests/test_k_means.py
@@ -16,7 +16,6 @@ from sklearn.utils.testing import assert_greater
 from sklearn.utils.testing import assert_less
 from sklearn.utils.testing import assert_warns
 from sklearn.utils.testing import if_safe_multiprocessing_with_blas
-from sklearn.utils.testing import if_not_mac_os
 from sklearn.utils.testing import assert_raise_message
 
 
@@ -287,14 +286,18 @@ def test_k_means_explicit_init_shape():
         msg = "does not match the number of features of the data"
         assert_raises_regex(ValueError, msg, km.fit, X)
         # for callable init
-        km = Class(n_init=1, init=lambda X_, k, random_state: X_[:, :2], n_clusters=len(X))
+        km = Class(n_init=1,
+                   init=lambda X_, k, random_state: X_[:, :2],
+                   n_clusters=len(X))
         assert_raises_regex(ValueError, msg, km.fit, X)
         # mismatch of number of clusters
         msg = "does not match the number of clusters"
         km = Class(n_init=1, init=X[:2, :], n_clusters=3)
         assert_raises_regex(ValueError, msg, km.fit, X)
         # for callable init
-        km = Class(n_init=1, init=lambda X_, k, random_state: X_[:2, :], n_clusters=3)
+        km = Class(n_init=1,
+                   init=lambda X_, k, random_state: X_[:2, :],
+                   n_clusters=3)
         assert_raises_regex(ValueError, msg, km.fit, X)
 
 
@@ -771,4 +774,122 @@ def test_x_squared_norms_init_centroids():
 def test_max_iter_error():
 
     km = KMeans(max_iter=-1)
-    assert_raise_message(ValueError, 'Number of iterations should be', km.fit, X)
+    assert_raise_message(ValueError, 'Number of iterations should be',
+                         km.fit, X)
+
+
+def test_kmeans_float32_64():
+    km = KMeans(n_init=1, random_state=11)
+
+    # float64 data
+    km.fit(X)
+    # dtype of cluster centers has to be the dtype of the input data
+    assert_equal(km.cluster_centers_.dtype, np.float64)
+    inertia64 = km.inertia_
+    X_new64 = km.transform(km.cluster_centers_)
+    pred64 = km.predict(X[0])
+
+    # float32 data
+    km.fit(np.float32(X))
+    # dtype of cluster centers has to be the dtype of the input data
+    assert_equal(km.cluster_centers_.dtype, np.float32)
+    inertia32 = km.inertia_
+    X_new32 = km.transform(km.cluster_centers_)
+    pred32 = km.predict(X[0])
+
+    # compare arrays with low precision since the difference between
+    # 32 and 64 bit sometimes makes a difference up to the 4th decimal place
+    assert_array_almost_equal(inertia32, inertia64, decimal=4)
+    assert_array_almost_equal(X_new32, X_new64, decimal=4)
+    # both predictions have to be the same and correspond to the correct label
+    assert_equal(pred32, pred64)
+    assert_equal(pred32, km.labels_[0])
+    assert_equal(pred64, km.labels_[0])
+
+    # float64 sparse data
+    km.fit(X_csr)
+    # dtype of cluster centers has to be the dtype of the input data
+    assert_equal(km.cluster_centers_.dtype, np.float64)
+    inertia64 = km.inertia_
+    X_new64 = km.transform(km.cluster_centers_)
+    pred64 = km.predict(X_csr[0])
+
+    # float32 sparse data
+    # Note: at the moment sparse data is always processed as float64 internally
+    km.fit(sp.csr_matrix(X_csr, dtype=np.float32))
+    assert_equal(km.cluster_centers_.dtype, np.float64)
+    inertia32 = km.inertia_
+    X_new32 = km.transform(km.cluster_centers_)
+    pred32 = km.predict(X_csr[0])
+
+    assert_array_almost_equal(inertia32, inertia64)
+    assert_array_almost_equal(X_new32, X_new64)
+    # both predictions have to be the same and correspond to the correct label
+    assert_equal(pred32, pred64)
+    assert_equal(pred32, km.labels_[0])
+    assert_equal(pred64, km.labels_[0])
+
+
+def test_mb_k_means_float32_64():
+    km = MiniBatchKMeans(n_init=1, random_state=30)
+
+    # float64 data
+    km.fit(X)
+    # dtype of cluster centers has to be the dtype of the input data
+    assert_equal(km.cluster_centers_.dtype, np.float64)
+    inertia64 = km.inertia_
+    X_new64 = km.transform(km.cluster_centers_)
+    pred64 = km.predict(X[0])
+    km.partial_fit(X[0:3])
+    # dtype of cluster centers has to stay the same after partial_fit
+    assert_equal(km.cluster_centers_.dtype, np.float64)
+
+    # float32 data
+    km.fit(np.float32(X))
+    # dtype of cluster centers has to be the dtype of the input data
+    assert_equal(km.cluster_centers_.dtype, np.float32)
+    inertia32 = km.inertia_
+    X_new32 = km.transform(km.cluster_centers_)
+    pred32 = km.predict(X[0])
+    km.partial_fit(X[0:3])
+    # dtype of cluster centers has to stay the same after partial_fit
+    assert_equal(km.cluster_centers_.dtype, np.float32)
+
+    # compare arrays with low precision since the difference between
+    # 32 and 64 bit sometimes makes a difference up to the 4th decimal place
+    assert_array_almost_equal(inertia32, inertia64, decimal=4)
+    assert_array_almost_equal(X_new32, X_new64, decimal=4)
+    # both predictions have to be the same and correspond to the correct label
+    assert_equal(pred32, pred64)
+    assert_equal(pred32, km.labels_[0])
+    assert_equal(pred64, km.labels_[0])
+
+    # float64 sparse data
+    km.fit(X_csr)
+    # dtype of cluster centers has to be the dtype of the input data
+    assert_equal(km.cluster_centers_.dtype, np.float64)
+    inertia64 = km.inertia_
+    X_new64 = km.transform(km.cluster_centers_)
+    pred64 = km.predict(X_csr[0])
+    km.partial_fit(X_csr[0:3])
+    # dtype of cluster centers has to stay the same after partial_fit
+    assert_equal(km.cluster_centers_.dtype, np.float64)
+
+    # float32 sparse data
+    # Note: at the moment sparse data is always processed as float64 internally
+    km.fit(sp.csr_matrix(X_csr, dtype=np.float32))
+    # dtype of cluster centers has to be always float64 (see Note above.)
+    assert_equal(km.cluster_centers_.dtype, np.float64)
+    inertia32 = km.inertia_
+    X_new32 = km.transform(km.cluster_centers_)
+    pred32 = km.predict(X_csr[0])
+    km.partial_fit(X_csr[0:3])
+    # dtype of cluster centers has to stay the same after partial_fit
+    assert_equal(km.cluster_centers_.dtype, np.float64)
+
+    assert_array_almost_equal(inertia32, inertia64)
+    assert_array_almost_equal(X_new32, X_new64)
+    # both predictions have to be the same and correspond to the correct label
+    assert_equal(pred32, pred64)
+    assert_equal(pred32, km.labels_[0])
+    assert_equal(pred64, km.labels_[0])